requirements.txt
requests>=2.32
python-dotenv>=1.0


pip install -r requirements.txt
